# openapi_generated.yaml
openapi: 3.0.3
info:
  title: Agri Clinic API
  version: 1.0.0
  description: "Production-ready REST API for the Agri Clinic SaaS platform.\n\n##
    Authentication\nAll protected endpoints require a **Bearer JWT** token.  \nObtain
    tokens via `POST /api/v1/auth/login/` then pass:  \n`Authorization: Bearer <access_token>`\n\n##
    Standard Response Envelope\n**Success:** `{ success: true, message: string, data:
    object }`  \n**Error:**   `{ success: false, message: string, errors: object,
    code: string }`\n\n## Pagination\nAll list endpoints return `{ count, next, previous,
    results }` under `data`.\nUse `?page=N&page_size=N` query params."
paths:
...
  /api/v1/dashboard/summary/:
    get:
      operationId: v1_dashboard_summary_retrieve
      description: 'Returns top-level KPIs: total visits, active employees, farmer
        coverage, today''s visit count, etc.'
      summary: Dashboard summary stats
      tags:
      - Dashboard
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccess'
          description: ''
...
  /api/v1/notifications/{id}/read/:
    post:
      operationId: v1_notifications_read_create
      description: Marks a single notification as read. Returns 404 if not found or
        already read.
      summary: Mark notification as read
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Notifications
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccess'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationNotFound'
          description: ''
...
  /api/v1/notifications/mark-all-read/:
    post:
      operationId: v1_notifications_mark_all_read_create
      description: Marks all of the current user's unread notifications as read.
      summary: Mark all notifications as read
      tags:
      - Notifications
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccess'
          description: ''
  /api/v1/notifications/unread-count/:
    get:
      operationId: v1_notifications_unread_count_retrieve
      description: Returns the count of unread notifications for the authenticated
        user.
      summary: Unread notification count
      tags:
      - Notifications
      security:
      - jwtAuth: []
      - cookieAuth: []
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Unread count
                content:
                  application/json:
                    example:
                      success: true
                      message: OK
                      data:
                        unread_count: 5
          description: ''
...
components:
  schemas:
...
    NotificationNotFound:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
        errors:
          type: object
          additionalProperties: {}
          default: {}
        code:
          type: string
      required:
      - code
      - message
...
    SimpleSuccess:
      type: object
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
        data:
          type: object
          additionalProperties: {}
          default: {}
      required:
      - message
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...